fix: tainted handshake.io and add large client hello test (#5208)

+++ tls/s2n_security_policies.c
@@ -1509,7 +1509,7 @@ bool s2n_ecc_is_extension_required(const struct s2n_security_policy *security_po
     if (cipher_preferences == NULL) {
         return false;
     }
-    for (uint8_t i = 0; i < cipher_preferences->count; i++) {
+    for (size_t i = 0; i < cipher_preferences->count; i++) {
         if (s2n_cipher_suite_requires_ecc_extension(cipher_preferences->suites[i])) {
             return true;
         }
@@ -1540,7 +1540,7 @@ bool s2n_pq_kem_is_extension_required(const struct s2n_security_policy *security
     if (cipher_preferences == NULL) {
         return false;
     }
-    for (uint8_t i = 0; i < cipher_preferences->count; i++) {
+    for (size_t i = 0; i < cipher_preferences->count; i++) {
         if (s2n_cipher_suite_requires_pq_extension(cipher_preferences->suites[i])) {
             return true;
         }
@@ -1569,7 +1569,7 @@ bool s2n_security_policy_supports_tls13(const struct s2n_security_policy *securi
         return false;
     }
 
-    for (uint8_t i = 0; i < cipher_preferences->count; i++) {
+    for (size_t i = 0; i < cipher_preferences->count; i++) {
         if (cipher_preferences->suites[i]->minimum_required_tls_version >= S2N_TLS13) {
             return true;
         }
